+++ src/main/java/com/example/index/generators/BigramIndexGenerator.java
@@ -53,7 +53,7 @@ public final class BigramIndexGenerator extends IndexGenerator<AnnotationEntry> 
             if (isFirstBatch) {
                 stmt.setInt(1, this.batchSize);
             } else {
-                stmt.setInt(1, lastProcessedEntry.getAnnotationId());
+                stmt.setLong(1, lastProcessedEntry.getAnnotationId());
                 stmt.setInt(2, this.batchSize);
             }
