+++ src/main/java/com/example/index/generators/NerIndexGenerator.java
@@ -61,7 +61,7 @@ public final class NerIndexGenerator extends IndexGenerator<AnnotationEntry> {
             if (lastProcessedEntry == null) {
                 stmt.setInt(1, batchSize);
             } else {
-                stmt.setInt(1, lastProcessedEntry.getAnnotationId());
+                stmt.setLong(1, lastProcessedEntry.getAnnotationId());
                 stmt.setInt(2, batchSize);
             }
             try (ResultSet rs = stmt.executeQuery()) {
